moved trade classes outside of Sapi namespace for shorter names
