+++ spec/factories/annual_report_uploads.rb
@@ -1,5 +1,5 @@
 FactoryGirl.define do
-  factory :annual_report_upload, class: Sapi::Trade::AnnualReportUpload do
+  factory :annual_report_upload, class: Trade::AnnualReportUpload do
     trading_country
     point_of_view 'E'
     epix_created_at { Time.now }
